requirements.txt
langchain
ipykernel
python-dotenv
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai 
chromadb
torch
sentence-transformers
langchain-huggingface
tqdm
langchain-ollama
faiss-cpu
langchain-chroma
langgraph 
cassio
tiktoken 
langchain-groq 
langchainhub
streamlit
langchain-core
numpy
bert_score
plotly.express
ragas 
datasets 
evaluate
ollama 
ollama-api 
langchain-ollama